move: Constants.Color to KikurageUI
